079_Two_by_Two AC 数据范围问题

+++ AtCoder/typical90/079_Two_by_Two.cpp
@@ -3,40 +3,39 @@
 // 按照一下顺序进行修改：（1，1）（1，2）（1，W），（2，1），（2，2），（2，W）
 // 到此为止，没有“最优化”的部分（都是必须进行的步骤，策略没有优化的空间）
 
-// todo 有一组数据是不对的
-
 #include<algorithm>
 #include<iostream>
+#include<cstdint>
 
-const int maxn = 101;
+const int64_t maxn = 101;
 
-int h,w,a[maxn][maxn],b[maxn][maxn];
+int64_t h,w,a[maxn][maxn],b[maxn][maxn];
 
 int main()
 {
     std::cin >> h >> w;
-    for(int i=1;i<=h;i+=1)
+    for(int64_t i=1;i<=h;i+=1)
     {
-        for(int j=1;j<=w;j+=1)
+        for(int64_t j=1;j<=w;j+=1)
         {
             std::cin >> a[i][j];
         }
     }
 
-    for(int i=1;i<=h;i+=1)
+    for(int64_t i=1;i<=h;i+=1)
     {
-        for(int j=1;j<=w;j+=1)
+        for(int64_t j=1;j<=w;j+=1)
         {
             std::cin >> b[i][j];
         }
     }
 
-    int ans = 0;
-    for(int i=1;i<h;i+=1)
+    int64_t ans = 0;
+    for(int64_t i=1;i<h;i+=1)
     {
-        for(int j=1;j<w;j+=1)
+        for(int64_t j=1;j<w;j+=1)
         {
-            int diff = b[i][j] - a[i][j];
+            int64_t diff = b[i][j] - a[i][j];
             a[i][j] += diff;
             a[i+1][j] += diff;
             a[i][j+1] += diff;
@@ -45,7 +44,7 @@ int main()
         }
     }
     
-    for(int i=1;i<=h;i+=1)
+    for(int64_t i=1;i<=h;i+=1)
     {
         if(a[i][w] != b[i][w])
         {
@@ -53,7 +52,7 @@ int main()
             return 0;
         }
     }
-    for(int j=1;j<=w;j+=1)
+    for(int64_t j=1;j<=w;j+=1)
     {
         if(a[h][j] != b[h][j])
         {
